+++ src/test/java/org/yorm/YormMySqlTest.java
@@ -10,9 +10,7 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.yorm.exception.YormException;
-import org.yorm.records.Company;
-import org.yorm.records.Invoice;
-import org.yorm.records.Person;
+import org.yorm.records.*;
 import org.yorm.util.DbType;
 import org.yorm.utils.TestConnectionFactory;
 
@@ -66,10 +64,10 @@ class YormMySqlTest {
     @Order(2)
     void saveCompany() throws YormException {
         Company company = new Company(0, "Hogwarts", "GB", LocalDate.of(1968, 2, 12), 154.1f, true);
-        int id = yorm.save(company);
+        long id = yorm.save(company);
         assertEquals(1, id);
         Company company2 = new Company(0, "Mordor", "ZZ", LocalDate.of(114, 11, 5), 0f, false);
-        int id2 = yorm.save(company2);
+        long id2 = yorm.save(company2);
         assertEquals(2, id2);
     }
 
@@ -89,7 +87,7 @@ class YormMySqlTest {
     @Order(4)
     void savePerson() throws YormException {
         Person person = new Person(0, "John", "john.doe@um.com", LocalDateTime.of(2022, 3, 22, 11, 14, 13), 1);
-        int idPerson = yorm.save(person);
+        long idPerson = yorm.save(person);
         assertEquals(1, idPerson);
         Person personWrong = new Person(0, "John", "john.doe@um.com", LocalDateTime.of(2022, 1, 15, 7, 53, 21), 0);
         assertThrows(YormException.class, () -> yorm.save(personWrong));
